refactor user-address.dto
