+++ src/user/user.service.ts
@@ -13,7 +13,7 @@ import { Profile } from './entities/profile.entity';
 import { OTPType } from 'src/user/entities/user-otp.entity';
 import { IsNull } from 'typeorm';
 import { UserAdress } from './entities/user-address.entity';
-import { CreateUserAddressDTO } from './dto/create-user-address.dto';
+import { CreateUserAddressDTO } from './dto/user-address.dto';
 import { UpdateUserDTO } from './dto/user-update.dto';
 import { ConfigService } from '@nestjs/config';
